+++ javascript/react/src/components/ArqueroTableHeader/ArqueroTableheader.types1.ts
