+++ javascript/react/src/components/ArqueroTableHeader/__tests__/TableCommands/TableCommands.story.tsx
@@ -6,7 +6,7 @@ import type { ComponentStory } from '@storybook/react'
 
 import { TableCommands } from '../../../TableCommands.js'
 import { ArqueroTableHeader } from '../../ArqueroTableHeader.js'
-import type { ArqueroTableHeaderProps } from '../../ArqueroTableheader.types.js'
+import type { ArqueroTableHeaderProps } from '../../ArqueroTableheader.types1.js'
 import { useParameters } from './TableCommands.hooks.js'
 
 export const TableCommandsStory: ComponentStory<typeof ArqueroTableHeader> = (
