+++ javascript/react/src/components/ArqueroTableHeader/index.ts
@@ -3,5 +3,5 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 export * from './ArqueroTableHeader.js'
-export * from './ArqueroTableheader.types.js'
+export * from './ArqueroTableheader.types1.js'
 export * from './commands/index.js'
